change file structure, move to es6 modules
