--- src/Constants.js
@@ -1,11 +1,0 @@
-module.exports = {
-    keys: {
-        backspace: 8,
-        enter: 13,
-        esc: 27,
-        left: 37,
-        up: 38,
-        right: 39,
-        down: 40
-    }
-};
