+++ src/view-header.js
@@ -1,4 +1,4 @@
-var React = require('react');
+import React from 'react'
 
 module.exports = React.createClass({
